refactor: split rds.ts to rds/ classes
