+++ apps/antalmanac/src/backend/lib/rds/sessions-rds.ts
@@ -1,0 +1,65 @@
+import type { db } from '@packages/db';
+import type * as schema from '@packages/db/src/schema';
+import { sessions, type Session } from '@packages/db/src/schema';
+import { and, eq, ExtractTablesWithRelations } from 'drizzle-orm';
+import type { PgTransaction, PgQueryResultHKT } from 'drizzle-orm/pg-core';
+
+type Transaction = PgTransaction<PgQueryResultHKT, typeof schema, ExtractTablesWithRelations<typeof schema>>;
+type DatabaseOrTransaction = Omit<typeof db, '$client'> | Transaction;
+
+export class SessionsRDS {
+    /**
+     * Retrieves the current session from the database using the provided refresh token.
+     */
+    static async getCurrentSession(db: DatabaseOrTransaction, refreshToken: string) {
+        return db
+            .select()
+            .from(sessions)
+            .where(eq(sessions.refreshToken, refreshToken))
+            .then((res) => res[0] ?? null);
+    }
+
+    /**
+     * Creates a new session for a user in the database.
+     */
+    static async createSession(tx: Transaction, userID: string): Promise<Session | null> {
+        return tx
+            .insert(sessions)
+            .values({
+                userId: userID,
+                expires: new Date(Date.now() + 30 * 24 * 60 * 60 * 1000), // 30 days
+            })
+            .returning()
+            .then((res) => res[0] ?? null);
+    }
+
+    /**
+     * Removes a session from the database for a given user and refresh token.
+     * If refreshToken is null, no action is taken.
+     */
+    static async removeSession(db: DatabaseOrTransaction, userId: string, refreshToken: string | null) {
+        if (refreshToken) {
+            await db.delete(sessions).where(and(eq(sessions.userId, userId), eq(sessions.refreshToken, refreshToken)));
+        }
+    }
+
+    /**
+     * Returns the existing session if the refresh token matches, otherwise creates a new one.
+     */
+    static async upsertSession(
+        db: DatabaseOrTransaction,
+        userId: string,
+        refreshToken?: string
+    ): Promise<Session | null> {
+        return db.transaction(async (tx) => {
+            const currentSession = await tx
+                .select()
+                .from(sessions)
+                .where(eq(sessions.refreshToken, refreshToken ?? ''))
+                .then((res) => res[0] ?? null);
+
+            if (currentSession) return currentSession;
+            return await SessionsRDS.createSession(tx, userId);
+        });
+    }
+}
